+++ packages/actor-init-sparql/test/ActorInitSparql-test.ts
@@ -2,13 +2,12 @@ import {ProxyHandlerStatic} from "@comunica/actor-http-proxy";
 import {ActorInit} from "@comunica/bus-init";
 import {Bindings, KEY_CONTEXT_QUERY_TIMESTAMP} from "@comunica/bus-query-operation";
 import {Bus, KEY_CONTEXT_LOG} from "@comunica/core";
-import {literal, namedNode, variable} from "@rdfjs/data-model";
-import {Factory, translate} from "sparqlalgebrajs";
+import {literal, variable} from "@rdfjs/data-model";
+import {translate} from "sparqlalgebrajs";
 import {PassThrough, Readable} from "stream";
 import {ActorInitSparql} from "../lib/ActorInitSparql";
 import {ActorInitSparql as ActorInitSparqlBrowser, KEY_CONTEXT_QUERYFORMAT} from "../lib/ActorInitSparql-browser";
-
-const FACTORY: Factory = new Factory();
+import Factory from "sparqlalgebrajs/lib/factory";
 
 describe('ActorInitSparqlBrowser', () => {
   it('should not allow invoking its run method', () => {
@@ -69,88 +68,6 @@ describe('ActorInitSparql', () => {
     });
   });
 
-  describe('#applyInitialBindings', () => {
-    it('should filter out project variables that are defined', () => {
-      const input = FACTORY.createProject(FACTORY.createBgp([
-        FACTORY.createPattern(namedNode('s'), namedNode('p'), namedNode('o'), namedNode('g')),
-      ]), [ variable('a'), variable('b'), variable('c') ]);
-      expect(ActorInitSparql.applyInitialBindings(input, Bindings({ '?b': literal('bl') }))).toEqual(
-        FACTORY.createProject(FACTORY.createBgp([
-          FACTORY.createPattern(namedNode('s'), namedNode('p'), namedNode('o'), namedNode('g')),
-        ]), [ variable('a'), variable('c') ]),
-      );
-    });
-
-    it('should not change a BGP without variables', () => {
-      const input = FACTORY.createBgp([
-        FACTORY.createPattern(namedNode('s'), namedNode('p'), namedNode('o'), namedNode('g')),
-      ]);
-      expect(ActorInitSparql.applyInitialBindings(input, Bindings({}))).toEqual(input);
-    });
-
-    it('should not change a BGP with unbound variables', () => {
-      const input = FACTORY.createBgp([
-        FACTORY.createPattern(variable('s'), variable('p'), variable('o'), variable('g')),
-      ]);
-      expect(ActorInitSparql.applyInitialBindings(input, Bindings({
-        '?g1': literal('gl'),
-        '?o1': literal('ol'),
-        '?p1': literal('pl'),
-        '?s1': literal('sl'),
-      }))).toEqual(input);
-    });
-
-    it('should change a BGP with bound variables', () => {
-      const input = FACTORY.createBgp([
-        FACTORY.createPattern(variable('s'), variable('p'), variable('o'), variable('g')),
-      ]);
-      expect(ActorInitSparql.applyInitialBindings(input, Bindings({
-        '?g': literal('gl'),
-        '?o': literal('ol'),
-        '?p': literal('pl'),
-        '?s': literal('sl'),
-      }))).toEqual(FACTORY.createBgp([
-        FACTORY.createPattern(literal('sl'), literal('pl'), literal('ol'), literal('gl')),
-      ]));
-    });
-
-    it('should not change a path expression without bound variables', () => {
-      const input = FACTORY.createPath(variable('s'), FACTORY.createNps([namedNode('p')]),
-        variable('o'), variable('g'));
-      expect(ActorInitSparql.applyInitialBindings(input, Bindings({
-        '?g1': literal('gl'),
-        '?o1': literal('ol'),
-        '?p1': literal('pl'),
-        '?s1': literal('sl'),
-      }))).toEqual(FACTORY.createPath(variable('s'), FACTORY.createNps([namedNode('p')]),
-        variable('o'), variable('g')));
-    });
-
-    it('should change a path expression with bound variables', () => {
-      const input = FACTORY.createPath(variable('s'), FACTORY.createNps([namedNode('p')]),
-        variable('o'), variable('g'));
-      expect(ActorInitSparql.applyInitialBindings(input, Bindings({
-        '?g': literal('gl'),
-        '?o': literal('ol'),
-        '?p': literal('pl'),
-        '?s': literal('sl'),
-      }))).toEqual(FACTORY.createPath(literal('sl'), FACTORY.createNps([namedNode('p')]),
-        literal('ol'), literal('gl')));
-    });
-
-    it('should correctly map describe operations', () => {
-      const input = FACTORY.createDescribe(FACTORY.createBgp([
-        FACTORY.createPattern(namedNode('s'), namedNode('p'), namedNode('o'), namedNode('g')),
-      ]), [ namedNode('a'), namedNode('b') ]);
-      expect(ActorInitSparql.applyInitialBindings(input, Bindings({}))).toEqual(input);
-    });
-
-    it('should correctly map values operations', () => {
-      const input = FACTORY.createValues([ variable('a') ], [{ '?a': namedNode('b') }]);
-      expect(ActorInitSparql.applyInitialBindings(input, Bindings({}))).toEqual(input);
-    });
-  });
-
   describe('An ActorInitSparql instance', () => {
     const hypermedia: string = "http://example.org/";
     const hypermedia2: string = "hypermedia@http://example.org/";
@@ -168,18 +85,31 @@ describe('ActorInitSparql', () => {
         input.push({ a: 'triple' });
         input.push(null);
       };
+      const factory = new Factory();
       mediatorQueryOperation.mediate = (action) => action.operation.query !== 'INVALID'
         ? Promise.resolve({ bindingsStream: input }) : Promise.reject(new Error('a'));
-      mediatorSparqlParse.mediate = (action) => Promise.resolve({
-        baseIRI: action.query.indexOf('BASE') >= 0 ? 'myBaseIRI' : null,
-        operation: action,
-      });
+      mediatorSparqlParse.mediate = (action) => action.query === 'INVALID'
+        ? Promise.resolve(action.query)
+        : Promise.resolve({
+          baseIRI: action.query.indexOf('BASE') >= 0 ? 'myBaseIRI' : null,
+          operation: factory.createProject(
+            factory.createBgp([
+              factory.createPattern(variable('s'), variable('p'), variable('o')),
+            ]),
+            [
+              variable('s'),
+              variable('p'),
+              variable('o')
+            ]
+          ),
+        });
       const defaultQueryInputFormat = 'sparql';
       actor = new ActorInitSparql(
         { bus, contextKeyShortcuts, defaultQueryInputFormat, logger, mediatorContextPreprocess,
           mediatorHttpInvalidate, mediatorOptimizeQueryOperation, mediatorQueryOperation,
           mediatorSparqlParse, mediatorSparqlSerialize,
-          mediatorSparqlSerializeMediaTypeCombiner: mediatorSparqlSerialize, name: 'actor' });
+          mediatorSparqlSerializeMediaTypeCombiner: mediatorSparqlSerialize,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: mediatorSparqlSerialize, name: 'actor' });
     });
 
     it('should test', () => {
@@ -226,7 +156,8 @@ describe('ActorInitSparql', () => {
         { bus, contextKeyShortcuts, defaultQueryInputFormat: 'sparql', logger, mediatorContextPreprocess,
           mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation, mediatorSparqlParse, mediatorSparqlSerialize,
-          mediatorSparqlSerializeMediaTypeCombiner: mediatorSparqlSerialize, name: 'actor', queryString });
+          mediatorSparqlSerializeMediaTypeCombiner: mediatorSparqlSerialize,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: mediatorSparqlSerialize, name: 'actor', queryString });
       const spy = jest.spyOn(actor, 'query');
       (await actor.run({ argv: [], env: {}, stdin: new PassThrough() }));
       return expect(spy.mock.calls[0][1][KEY_CONTEXT_QUERYFORMAT]).toEqual('sparql');
@@ -237,6 +168,7 @@ describe('ActorInitSparql', () => {
         { bus, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation, mediatorSparqlParse,
           mediatorSparqlSerialize, mediatorSparqlSerializeMediaTypeCombiner: mediatorSparqlSerialize,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: mediatorSparqlSerialize,
           name: 'actor', queryString });
       const spy = jest.spyOn(actor, 'query');
       (await actor.run({ argv: [ '-i', 'graphql' ], env: {}, stdin: new PassThrough() }));
@@ -250,7 +182,8 @@ describe('ActorInitSparql', () => {
       actor = new ActorInitSparql(
         { bus, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation, mediatorSparqlParse,
-          mediatorSparqlSerialize: med, mediatorSparqlSerializeMediaTypeCombiner: med, name: 'actor', queryString });
+          mediatorSparqlSerialize: med, mediatorSparqlSerializeMediaTypeCombiner: med,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: med, name: 'actor', queryString });
       return expect((await actor.run({ argv: [ '-t', 'testtype' ], env: {}, stdin: new PassThrough() })).stdout)
         .toEqual('testtype');
     });
@@ -265,7 +198,8 @@ describe('ActorInitSparql', () => {
       actor = new ActorInitSparql(
         { bus, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation: m1, mediatorSparqlParse,
-          mediatorSparqlSerialize: m2, mediatorSparqlSerializeMediaTypeCombiner: m2, name: 'actor', queryString });
+          mediatorSparqlSerialize: m2, mediatorSparqlSerializeMediaTypeCombiner: m2,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: m2, name: 'actor', queryString });
       return expect((await actor.run({ argv: [], env: {}, stdin: new PassThrough() })).stdout)
         .toEqual('application/json');
     });
@@ -280,7 +214,8 @@ describe('ActorInitSparql', () => {
       actor = new ActorInitSparql(
         { bus, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation: m1, mediatorSparqlParse,
-          mediatorSparqlSerialize: m2, mediatorSparqlSerializeMediaTypeCombiner: m2, name: 'actor', queryString });
+          mediatorSparqlSerialize: m2, mediatorSparqlSerializeMediaTypeCombiner: m2,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: m2, name: 'actor', queryString });
       return expect((await actor.run({ argv: [], env: {}, stdin: new PassThrough() })).stdout)
         .toEqual('application/trig');
     });
@@ -295,7 +230,8 @@ describe('ActorInitSparql', () => {
       actor = new ActorInitSparql(
         { bus, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation: m1, mediatorSparqlParse,
-          mediatorSparqlSerialize: m2, mediatorSparqlSerializeMediaTypeCombiner: m2, name: 'actor', queryString });
+          mediatorSparqlSerialize: m2, mediatorSparqlSerializeMediaTypeCombiner: m2,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: m2, name: 'actor', queryString });
       return expect((await actor.run({ argv: [], env: {}, stdin: new PassThrough() })).stdout)
         .toEqual('simple');
     });
@@ -305,6 +241,7 @@ describe('ActorInitSparql', () => {
         { bus, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation, mediatorSparqlParse,
           mediatorSparqlSerialize, mediatorSparqlSerializeMediaTypeCombiner: mediatorSparqlSerialize,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: mediatorSparqlSerialize,
           name: 'actor', queryString });
       return actor.run({ argv: [ ], env: {}, stdin: new PassThrough() })
         .then((result) => {
@@ -320,7 +257,8 @@ describe('ActorInitSparql', () => {
         { bus, context, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation,
           mediatorSparqlParse, mediatorSparqlSerialize,
-          mediatorSparqlSerializeMediaTypeCombiner: mediatorSparqlSerialize, name: 'actor', queryString });
+          mediatorSparqlSerializeMediaTypeCombiner: mediatorSparqlSerialize,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: mediatorSparqlSerialize, name: 'actor', queryString });
       return actor.run({ argv: [ ], env: {}, stdin: new PassThrough() })
         .then((result) => {
           return new Promise((resolve, reject) => {
@@ -335,7 +273,8 @@ describe('ActorInitSparql', () => {
         { bus, context, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation,
           mediatorSparqlParse, mediatorSparqlSerialize,
-          mediatorSparqlSerializeMediaTypeCombiner: mediatorSparqlSerialize, name: 'actor' });
+          mediatorSparqlSerializeMediaTypeCombiner: mediatorSparqlSerialize,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: mediatorSparqlSerialize, name: 'actor' });
       return expect(actor.run({ argv: [ ], env: {}, stdin: new PassThrough() })).resolves
         .toHaveProperty('stderr');
     });
@@ -345,7 +284,8 @@ describe('ActorInitSparql', () => {
         { bus, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation, mediatorSparqlParse,
           mediatorSparqlSerialize,
-          mediatorSparqlSerializeMediaTypeCombiner: mediatorSparqlSerialize, name: 'actor', queryString: null });
+          mediatorSparqlSerializeMediaTypeCombiner: mediatorSparqlSerialize,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: mediatorSparqlSerialize, name: 'actor', queryString: null });
       return expect(actor.run({ argv: [ ], env: {}, stdin: new PassThrough() })).resolves
         .toHaveProperty('stderr');
     });
@@ -521,7 +461,8 @@ describe('ActorInitSparql', () => {
       actor = new ActorInitSparql(
         { bus, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation, mediatorSparqlParse,
-          mediatorSparqlSerialize: med, mediatorSparqlSerializeMediaTypeCombiner: med, name: 'actor', queryString });
+          mediatorSparqlSerialize: med, mediatorSparqlSerializeMediaTypeCombiner: med,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: med, name: 'actor', queryString });
       return expect((await actor.run({
         argv: [ hypermedia, queryString, '-d', dt.toISOString() ], env: {},
         stdin: new PassThrough(),
@@ -535,7 +476,8 @@ describe('ActorInitSparql', () => {
       actor = new ActorInitSparql(
         { bus, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation, mediatorSparqlParse,
-          mediatorSparqlSerialize: med, mediatorSparqlSerializeMediaTypeCombiner: med, name: 'actor', queryString });
+          mediatorSparqlSerialize: med, mediatorSparqlSerializeMediaTypeCombiner: med,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: med, name: 'actor', queryString });
       return expect((await actor.run({
         argv: [ hypermedia, queryString, '-l', 'warn' ], env: {},
         stdin: new PassThrough(),
@@ -568,7 +510,8 @@ graph <exists02.ttl> {
       actor = new ActorInitSparql(
         { bus, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation, mediatorSparqlParse,
-          mediatorSparqlSerialize: med, mediatorSparqlSerializeMediaTypeCombiner: med, name: 'actor', queryString });
+          mediatorSparqlSerialize: med, mediatorSparqlSerializeMediaTypeCombiner: med,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: med, name: 'actor', queryString });
       const spy = jest.spyOn(actor, 'query');
       expect((await actor.run({
         argv: [ hypermedia, queryString, '-p', proxy ], env: {},
@@ -585,13 +528,29 @@ graph <exists02.ttl> {
       actor = new ActorInitSparql(
         { bus, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
           mediatorOptimizeQueryOperation, mediatorQueryOperation, mediatorSparqlParse,
-          mediatorSparqlSerialize: med, mediatorSparqlSerializeMediaTypeCombiner: med, name: 'actor', queryString });
+          mediatorSparqlSerialize: med, mediatorSparqlSerializeMediaTypeCombiner: med,
+          mediatorSparqlSerializeMediaTypeFormatCombiner: med, name: 'actor', queryString });
       const spy = jest.spyOn(actor, 'query');
       expect((await actor.run({
         argv: [ hypermedia, queryString, '--lenient' ], env: {},
         stdin: new PassThrough(),
       }))).toBeTruthy();
       expect(spy.mock.calls[0][1]['@comunica/actor-init-sparql:lenient']).toBeTruthy();
+    });
+
+    describe('getResultMediaTypeFormats', () => {
+      it('should return the media type formats', () => {
+        const med: any = {
+          mediate: (arg) => Promise.resolve({ mediaTypeFormats: { data: 'DATA' } }),
+        };
+        actor = new ActorInitSparql(
+          { bus, contextKeyShortcuts, logger, mediatorContextPreprocess, mediatorHttpInvalidate,
+            mediatorOptimizeQueryOperation, mediatorQueryOperation, mediatorSparqlParse,
+            mediatorSparqlSerialize: med, mediatorSparqlSerializeMediaTypeCombiner: med,
+            mediatorSparqlSerializeMediaTypeFormatCombiner: med, name: 'actor', queryString });
+        return expect(actor.getResultMediaTypeFormats(null))
+          .resolves.toEqual({ data: 'DATA' });
+      });
     });
 
   });